ci: Add new robot test 'button-left-click'
This test verifies button-left clicks work correctly.

A 'click' event is emitted then when the cursor position at which the
'mousedown' event and the 'mouseup' event happen in mostly the same, or
within a certain threshold.

When the 'mousedown' and 'mouseup' events occur at very different cursor
position, a 'click' event should not be emitted.

Maintenance-type: ci